+++ src/pytest_inmanta_lsm/plugin.py
@@ -15,6 +15,7 @@ import shutil
 import tempfile
 import textwrap
 import time
+import urllib.parse
 import uuid
 from typing import Any, Dict, Generator, Iterator, Optional, Tuple, Union
 from uuid import UUID
@@ -24,6 +25,7 @@ import pytest
 import pytest_inmanta.plugin
 import pytest_inmanta.test_parameter
 from inmanta import module
+from inmanta.data import model
 from packaging import version
 from pytest_inmanta.plugin import Project
 from pytest_inmanta.test_parameter import (
@@ -415,6 +417,33 @@ def remote_orchestrator_shared(
     tested is v2, it is installed in editable mode, as it is required to send it to the remote
     orchestrator.
     """
+    # Check if the environment is protected.  If it is, abort now with a clear error message
+    settings = remote_orchestrator_access.sync_request(
+        "environment_setting_get",
+        model.EnvironmentSettingsReponse,
+        tid=remote_orchestrator_access.environment,
+        id="protected_environment",
+    )
+    if settings.settings["protected_environment"]:
+        # Create a url that points to the settings page in the web console, and highlights the
+        # relevant setting
+        settings_url = urllib.parse.urlunsplit(
+            [
+                *remote_orchestrator_access.url_split[:2],  # schema, netloc
+                "/console/settings",  # path
+                urllib.parse.urlencode(
+                    {
+                        "env": remote_orchestrator_access.environment,
+                        "state.Settings.tab": "Configuration",
+                    },
+                ),  # query
+                # Highlight the setting on the web console page
+                # https://web.dev/articles/text-fragments#start
+                ":~:text=protected_environment",  # fragment
+            ],
+        )
+        raise RuntimeError(f"Environment is protected, it can't be used with pytest-inmanta-lsm: {settings_url}")
+
     # no need to do anything if this version of inmanta does not support v2 modules
     if hasattr(module, "ModuleV2"):
         verify_v2_editable_install()