+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v 3.10.0 (?)
 Changes in this release:
+- Make sure protected environments can never be cleaned up by pytest-inmanta-lsm fixtures.
 
 # v 3.9.0 (2024-10-29)
 Changes in this release: